Added retry mechanism for all remaining clipboard tests
