+++ src/Microsoft.VisualBasic/tests/UnitTests/Microsoft/VisualBasic/Devices/ComputerTests.cs
@@ -5,6 +5,7 @@ namespace Microsoft.VisualBasic.Devices.Tests;
 
 [Collection("Sequential")]
 [CollectionDefinition("Sequential", DisableParallelization = true)]
+[UISettings(MaxAttempts = 3)] // Try up to 3 times before failing.
 public class ComputerTests
 {
     [Fact]
