+++ src/Microsoft.VisualBasic/tests/UnitTests/Microsoft/VisualBasic/MyServices/ClipboardProxyTests.cs
@@ -13,6 +13,7 @@ namespace Microsoft.VisualBasic.MyServices.Tests;
 // Each registered Clipboard format is an OS singleton,
 // and we should not run this test at the same time as other tests using the same format.
 [Collection("Sequential")]
+[UISettings(MaxAttempts = 3)] // Try up to 3 times before failing.
 public class ClipboardProxyTests
 {
     private static string GetUniqueText() => Guid.NewGuid().ToString("D");
