+++ src/System.Windows.Forms/tests/UnitTests/System/Windows/Forms/ClipboardTests.cs
@@ -17,6 +17,7 @@ namespace System.Windows.Forms.Tests;
 // Note: each registered Clipboard format is an OS singleton
 // and we should not run this test at the same time as other tests using the same format.
 [Collection("Sequential")]
+[UISettings(MaxAttempts = 3)] // Try up to 3 times before failing.
 public class ClipboardTests
 {
     [WinFormsFact]
