+++ src/System.Windows.Forms/tests/UnitTests/System/Windows/Forms/DataObjectTests.ClipboardTests.cs
@@ -11,6 +11,7 @@ public partial class DataObjectTests
 {
     [Collection("Sequential")] // Each registered Clipboard format is an OS singleton,
                                // and we should not run this test at the same time as other tests using the same format.
+    [UISettings(MaxAttempts = 3)] // Try up to 3 times before failing.
     public class ClipboardTests
     {
         public static TheoryData<string, bool> GetData_StringBool_TheoryData()
