+++ src/System.Windows.Forms/tests/UnitTests/System/Windows/Forms/DragDropFormatTests.cs
@@ -17,6 +17,7 @@ namespace System.Windows.Forms.Tests;
 
 [Collection("Sequential")] // Each registered Clipboard format is an OS singleton,
                            // and we should not run this test at the same time as other tests using the same format.
+[UISettings(MaxAttempts = 3)] // Try up to 3 times before failing.
 public unsafe class DragDropFormatTests
 {
     public static IEnumerable<object[]> DragDropFormat_TestData()
